Split the implementation from daniel-dewar into separate files.
Cleaned up the main.cpp file.
