--- HashNode.h
@@ -1,24 +1,0 @@
-//
-// Created by rroig on 4/18/2024.
-//
-
-#ifndef DATASTRUCTURESFINALPROJECT_HASHNODE_H
-#define DATASTRUCTURESFINALPROJECT_HASHNODE_H
-
-// TODO: CHANGE THIS CLASS AROUND
-
-template <typename K, typename V>
-class HashNode {
-public:
-    V value;
-    K key;
-
-    // Constructor of hashnode
-    HashNode(K key, V value);
-
-};
-
-
-
-
-#endif //DATASTRUCTURESFINALPROJECT_HASHNODE_H
