+++ MiniHashNode.cpp
@@ -1,0 +1,12 @@
+//
+// Created by rroig on 4/18/2024.
+//
+
+#include "MiniHashNode.h"
+
+
+MiniHashNode::MiniHashNode(string KEY,  AirportNode* DATA)
+{
+Key = std::move(KEY);
+Data = DATA;
+}
